requirements.txt
PyMuPDF>=1.24.0
python-pptx>=0.6.23
google-generativeai>=0.8.0
markdown>=3.6
weasyprint>=62.0
python-dotenv>=1.0.0
